Depend on exact build-time ggml version
Upstream ships llama.cpp with a specific version of ggml. We have no
guarantees that any version earlier or later than that, in fact it's
common for newer versions to break something.

So going forward, we ship llama.cpp and ggml in tandem, with ggml being
updated first, and llama.cpp depending on the exact version used at
build-time.